requirements.txt
asgiref==3.11.1
Django==6.0.3
django-rest-passwordreset==1.5.0
djangorestframework==3.17.1
sqlparse==0.5.5
tzdata==2025.3
